+++ src/Stratis.Bitcoin.Features.Consensus/ConsensusManager.cs
@@ -8,11 +8,8 @@ using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.Features.Consensus
 {
-    public class ConsensusManager : IBlockDownloadState, INetworkDifficulty, IGetUnspentTransaction
+    public class ConsensusManager : INetworkDifficulty, IGetUnspentTransaction
     {
-        /// <summary>Provider of block header hash checkpoints.</summary>
-        private readonly ICheckpoints checkpoints;
-
         public ConsensusLoop ConsensusLoop { get; private set; }
 
         public IDateTimeProvider DateTimeProvider { get; private set; }
@@ -23,7 +20,7 @@ namespace Stratis.Bitcoin.Features.Consensus
 
         public PowConsensusValidator ConsensusValidator { get; private set; }
 
-        public ConsensusManager(ICheckpoints checkpoints, ConsensusLoop consensusLoop = null, IDateTimeProvider dateTimeProvider = null, NodeSettings nodeSettings = null, Network network = null,
+        public ConsensusManager(ConsensusLoop consensusLoop = null, IDateTimeProvider dateTimeProvider = null, NodeSettings nodeSettings = null, Network network = null,
             PowConsensusValidator consensusValidator = null)
         {
             this.ConsensusLoop = consensusLoop;
@@ -31,31 +28,6 @@ namespace Stratis.Bitcoin.Features.Consensus
             this.NodeSettings = nodeSettings;
             this.Network = network;
             this.ConsensusValidator = consensusValidator;
-            this.checkpoints = checkpoints;
-        }
-
-        /// <summary>
-        /// Checks whether the node is currently in the process of initial block download.
-        /// </summary>
-        /// <returns><c>true</c> if the node is currently doing IBD, <c>false</c> otherwise.</returns>
-        public bool IsInitialBlockDownload()
-        {
-            if (this.ConsensusLoop == null)
-                return false;
-
-            if (this.ConsensusLoop.Tip == null)
-                return true;
-
-            if (this.checkpoints.GetLastCheckpointHeight() > this.ConsensusLoop.Tip.Height)
-                return true;
-
-            if (this.ConsensusLoop.Tip.ChainWork < (this.Network.Consensus.MinimumChainWork ?? uint256.Zero))
-                return true;
-
-            if (this.ConsensusLoop.Tip.Header.BlockTime.ToUnixTimeSeconds() < (this.DateTimeProvider.GetTime() - this.NodeSettings.MaxTipAge))
-                return true;
-
-            return false;
         }
 
         public Target GetNetworkDifficulty()
