--- src/Stratis.Bitcoin/Interfaces/IBlockDownloadState.cs
@@ -1,7 +1,0 @@
-﻿namespace Stratis.Bitcoin.Interfaces
-{
-    public interface IBlockDownloadState
-    {
-        bool IsInitialBlockDownload();
-    }
-}
